+++ application.go
@@ -1,18 +1,22 @@
-package gwf
+package foundation
 
 import (
 	"fmt"
+	"github.com/RobyFerro/go-web-framework/kernel"
+	"github.com/RobyFerro/go-web-framework/register"
 	"github.com/common-nighthawk/go-figure"
+	"log"
 	"os"
 	"reflect"
 )
 
 // Start method will start the main Go-Web HTTP Server.
-func Start(args []string, cm CommandRegister, c ControllerRegister, s ServiceRegister, mw interface{}, m ModelRegister) {
+func Start(args []string, cm register.CommandRegister, c register.ControllerRegister, s register.ServiceRegister, mw interface{},
+	m register.ModelRegister) {
 	printCLIHeader()
 	registerBaseEntities(c, m, s, cm, mw)
 
-	cmd := Commands.List[args[0]]
+	cmd := kernel.Commands.List[args[0]]
 	if cmd == nil {
 		fmt.Println("Command not found!")
 		os.Exit(1)
@@ -26,18 +30,19 @@ func Start(args []string, cm CommandRegister, c ControllerRegister, s ServiceReg
 
 	// Build service container.
 	// This container will used to invoke the requested command.
-	container := BuildContainer()
+	container := kernel.BuildContainer()
 	if err := container.Invoke(rc.MethodByName("Run").Interface()); err != nil {
-		ProcessError(err)
+		log.Fatal(err)
 	}
 }
 
 // Register base entities in Go-Web kernel
 // This method will register: Controllers, Models, CLI commands, Services and middleware
-func registerBaseEntities(c ControllerRegister, m ModelRegister, s ServiceRegister, cm CommandRegister, mw interface{}) {
-	Controllers = c
-	Middleware = mw
-	Models = m
+func registerBaseEntities(c register.ControllerRegister, m register.ModelRegister, s register.ServiceRegister,
+	cm register.CommandRegister, mw interface{}) {
+	kernel.Controllers = c
+	kernel.Middleware = mw
+	kernel.Models = m
 
 	mergeCommands(cm)
 	bindServices(s.List)
@@ -45,9 +50,8 @@ func registerBaseEntities(c ControllerRegister, m ModelRegister, s ServiceRegist
 
 // Merge custom services with defaults
 func bindServices(services []interface{}) {
-
 	for _, s := range services {
-		Services.List = append(Services.List, s)
+		kernel.Services.List = append(kernel.Services.List, s)
 	}
 }
 
@@ -60,8 +64,8 @@ func printCLIHeader() {
 }
 
 // MergeCommands will merge system command with customs
-func mergeCommands(commands CommandRegister) {
+func mergeCommands(commands register.CommandRegister) {
 	for i, c := range commands.List {
-		Commands.List[i] = c
+		kernel.Commands.List[i] = c
 	}
 }