+++ cli/database_seed.go
@@ -1,7 +1,8 @@
-package gwf
+package cli
 
 import (
 	"fmt"
+	"github.com/RobyFerro/go-web-framework/register"
 	"reflect"
 	"strings"
 
@@ -23,7 +24,7 @@ func (c *Seeder) Register() {
 
 // Run this command
 // Todo: Improve this method to run a single seeder
-func (c *Seeder) Run(db *gorm.DB, models ModelRegister) {
+func (c *Seeder) Run(db *gorm.DB, models register.ModelRegister) {
 	fmt.Println("Execute database seeding...")
 	if len(c.Args) > 0 {
 		extractSpecificModel(c.Args, &models.List)