+++ cli/migration_rollback.go
@@ -1,8 +1,9 @@
-package gwf
+package cli
 
 import (
 	"fmt"
 	"io/ioutil"
+	"log"
 	"strconv"
 	"strings"
 
@@ -31,7 +32,7 @@ func (c *MigrateRollback) Run(db *gorm.DB) {
 	for i := 0; i < step; i++ {
 		var migrations []migration
 		if err := db.Order("created_at", true).Where("batch LIKE ?", batch).Find(&migrations).Error; err != nil {
-			ProcessError(err)
+			log.Fatal(err)
 		}
 
 		// Execute given rollback
@@ -48,13 +49,13 @@ func rollbackMigrations(migrations []migration, db *gorm.DB) {
 		fmt.Printf("\nRolling back '%s' migration...\n", rollbackFile)
 
 		if payload, err := ioutil.ReadFile(rollbackFile); err != nil {
-			ProcessError(err)
+			log.Fatal(err)
 		} else {
 			db.Exec(string(payload)).Row()
 		}
 
 		if err := db.Unscoped().Delete(&m).Error; err != nil {
-			ProcessError(err)
+			log.Fatal(err)
 		}
 
 		fmt.Printf("\nSuccess! %s has been rolled back!", rollbackFile)