+++ cli/model_create.go
@@ -1,8 +1,10 @@
-package gwf
+package cli
 
 import (
 	"fmt"
+	"github.com/RobyFerro/go-web-framework/tool"
 	"io/ioutil"
+	"log"
 	"path"
 	"path/filepath"
 	"regexp"
@@ -32,10 +34,10 @@ func (c *ModelCreate) Run() {
 	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "raw/model.raw"))
 
 	cContent := strings.ReplaceAll(string(input), "@@TMP@@", cName)
-	cFile := fmt.Sprintf("%s/%s.go", GetDynamicPath("database/model"), strings.ToLower(c.Args))
+	cFile := fmt.Sprintf("%s/%s.go", tool.GetDynamicPath("database/model"), strings.ToLower(c.Args))
 
 	if err := ioutil.WriteFile(cFile, []byte(cContent), 0755); err != nil {
-		ProcessError(err)
+		log.Fatal(err)
 	}
 
 	re := regexp.MustCompile("(model\\.[A-Za-z]\\w+( *){},(\\n*)(\\t*| *))")