+++ kernel/di.go
@@ -1,7 +1,13 @@
-package gwf
+package kernel
 
 import (
+	"github.com/RobyFerro/go-web-framework/register"
 	"go.uber.org/dig"
+)
+
+// Container will provide access to the global Service Container
+var (
+	Container *dig.Container
 )
 
 // BuildContainer provide the global service container
@@ -10,7 +16,7 @@ func BuildContainer() *dig.Container {
 
 	for _, s := range Services.List {
 		if err := container.Provide(s); err != nil {
-			ProcessError(err)
+
 		}
 	}
 
@@ -22,15 +28,15 @@ func BuildContainer() *dig.Container {
 
 // Inject base entities: controllers, models, commands in service container
 func injectBasicEntities(sc *dig.Container) {
-	_ = sc.Provide(func() ControllerRegister {
+	_ = sc.Provide(func() register.ControllerRegister {
 		return Controllers
 	})
 
-	_ = sc.Provide(func() CommandRegister {
+	_ = sc.Provide(func() register.CommandRegister {
 		return Commands
 	})
 
-	_ = sc.Provide(func() ModelRegister {
+	_ = sc.Provide(func() register.ModelRegister {
 		return Models
 	})
 }