+++ kernel/router.go
@@ -1,52 +1,76 @@
-package gwf
+package kernel
 
 import (
 	"fmt"
 	"github.com/gorilla/mux"
-	"go.uber.org/dig"
+	"github.com/RobyFerro/go-web-framework/tool"
+	"log"
 	"net/http"
 	"reflect"
 	"strings"
+	"sync"
 )
 
-var (
-	// BC is used to declare base controller
-	BC BaseController
-	// Container will provide access to the global Service Container
-	Container *dig.Container
-	// Controllers will handle every application controller
-	Controllers ControllerRegister
-	// Middleware will handle every application middleware
-	Middleware interface{}
-	// Models will handle every application middleware
-	Models ModelRegister
-	// Services will handle every application service
-	Services = ServiceRegister{
-		List: []interface{}{
-			Configuration,
-			CreateSessionStore,
-			GetHttpServer,
-			WebRouter,
-		},
-	}
-	// Commands will export all registered commands
-	// The following map of interfaces expose all available method that can be used by Go-Web CLI tool.
-	// The map index determines the command that you've to run to for use the relative method.
-	// Example: './goweb migration:up' will run '&command.MigrationUp{}' command.
-	Commands = CommandRegister{
-		List: map[string]interface{}{
-			"database:seed": &Seeder{},
-			"server:daemon": &ServerDaemon{},
-			"server:run":    &ServerRun{},
-			"show:commands": &ShowCommands{},
-			// Here is where you've to register your custom controller
-		},
-	}
-)
+// Route structure is used to decode all route presents into routing.yml file.
+type Route struct {
+	Path        string   `yaml:"path"`
+	Action      string   `yaml:"action"`
+	Method      string   `yaml:"method"`
+	Description string   `yaml:"description"`
+	Middleware  []string `yaml:"middleware"`
+	Prefix      string   `yaml:"prefix"`
+}
 
-// Handle single path parsing.
+// Group structure used to decode all groups presents into the routing.yml file.
+type Group struct {
+	Prefix     string `yaml:"prefix"`
+	Routes     map[string]Route
+	Middleware []string
+}
+
+// Router structure of web router.
+type Router struct {
+	Routes map[string]Route `yaml:"routes"`
+	Groups map[string]Group `yaml:"groups"`
+}
+
+// WebRouter parses routing structures and set every route.
+// Return a Gorilla Mux router instance with all routes indicated in router.yml file.
+func WebRouter() *mux.Router {
+	var wg sync.WaitGroup
+
+	wg.Add(3)
+
+	routes, err := RetrieveRoutingConf()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	router := mux.NewRouter()
+
+	go func() {
+		HandleSingleRoute(routes.Routes, router)
+		wg.Done()
+	}()
+
+	go func() {
+		HandleGroups(routes.Groups, router)
+		wg.Done()
+	}()
+
+	go func() {
+		GiveAccessToPublicFolder(router)
+		wg.Done()
+	}()
+
+	wg.Wait()
+
+	return router
+}
+
+// HandleSingleRoute handles single path parsing.
 // This method it's used to parse every single path. If middleware is present a sub-router with will be created
-func handleSingleRoute(routes map[string]Route, router *mux.Router) {
+func HandleSingleRoute(routes map[string]Route, router *mux.Router) {
 	for _, route := range routes {
 		hasMiddleware := len(route.Middleware) > 0
 		directive := strings.Split(route.Action, "@")
@@ -56,7 +80,7 @@ func handleSingleRoute(routes map[string]Route, router *mux.Router) {
 				cc := GetControllerInterface(directive, writer, request)
 				method := reflect.ValueOf(cc).MethodByName(directive[1])
 				if err := Container.Invoke(method.Interface()); err != nil {
-					ProcessError(err)
+					log.Fatal(err)
 				}
 			}).Methods(route.Method)
 
@@ -67,15 +91,15 @@ func handleSingleRoute(routes map[string]Route, router *mux.Router) {
 				cc := GetControllerInterface(directive, writer, request)
 				method := reflect.ValueOf(cc).MethodByName(directive[1])
 				if err := Container.Invoke(method.Interface()); err != nil {
-					ProcessError(err)
+					log.Fatal(err)
 				}
 			}).Methods(route.Method)
 		}
 	}
 }
 
-// Parse route groups.
-func handleGroups(groups map[string]Group, router *mux.Router) {
+// HandleGroups parses route groups.
+func HandleGroups(groups map[string]Group, router *mux.Router) {
 	for _, group := range groups {
 		subRouter := router.PathPrefix(group.Prefix).Subrouter()
 
@@ -88,7 +112,7 @@ func handleGroups(groups map[string]Group, router *mux.Router) {
 					cc := GetControllerInterface(directive, writer, request)
 					method := reflect.ValueOf(cc).MethodByName(directive[1])
 					if err := Container.Invoke(method.Interface()); err != nil {
-						ProcessError(err)
+						log.Fatal(err)
 					}
 				}).Methods(route.Method)
 
@@ -99,7 +123,7 @@ func handleGroups(groups map[string]Group, router *mux.Router) {
 					cc := GetControllerInterface(directive, writer, request)
 					method := reflect.ValueOf(cc).MethodByName(directive[1])
 					if err := Container.Invoke(method.Interface()); err != nil {
-						ProcessError(err)
+						log.Fatal(err)
 					}
 				}).Methods(route.Method)
 			}
@@ -109,26 +133,10 @@ func handleGroups(groups map[string]Group, router *mux.Router) {
 	}
 }
 
-// Parse list of middleware and get an array of []mux.Middleware func
-// Required by Gorilla Mux
-func parseMiddleware(mwList []string, middleware interface{}) []mux.MiddlewareFunc {
-	var midFunc []mux.MiddlewareFunc
-
-	for _, mw := range mwList {
-		m := reflect.ValueOf(middleware)
-		method := m.MethodByName(mw)
-
-		callable := method.Interface().(func(handler http.Handler) http.Handler)
-		midFunc = append(midFunc, callable)
-	}
-
-	return midFunc
-}
-
-// Give access to public folder. With the /public prefix you can access to all of your assets.
+// GiveAccessToPublicFolder gives access to public folder. With the /public prefix you can access to all of your assets.
 // This is mandatory to access to public files (.js, .css, images, etc...).
-func giveAccessToPublicFolder(router *mux.Router) {
-	publicDirectory := http.Dir(GetDynamicPath("public"))
+func GiveAccessToPublicFolder(router *mux.Router) {
+	publicDirectory := http.Dir(tool.GetDynamicPath("public"))
 	router.PathPrefix("/public/").Handler(http.StripPrefix("/public/", http.FileServer(publicDirectory)))
 }
 
@@ -146,29 +154,4 @@ func GetControllerInterface(directive []string, w http.ResponseWriter, r *http.R
 	}
 
 	return result
-}
-
-// Parse a controller instance and implement it with the current base controller.
-// This operation will give you access to all basic controller properties.
-func registerBaseController(res http.ResponseWriter, req *http.Request, controller *interface{}) *interface{} {
-	if err := setBaseController(res, req); err != nil {
-		ProcessError(err)
-	}
-
-	c := reflect.ValueOf(*controller).Elem().FieldByName("BaseController")
-	c.Set(reflect.ValueOf(BC))
-
-	return controller
-}
-
-// Setting up the base controller instance (defined in conf.go) with the properties/method defined in the Service Container.
-// Here you can implement the BaseController content.
-// Remember to update even the structure (app/http/controller/controller.go)
-func setBaseController(res http.ResponseWriter, req *http.Request) error {
-	BC = BaseController{
-		Response: res,
-		Request:  req,
-	}
-
-	return nil
 }